Moved sendBundle() into v*/peer.go
