+++ waku/v0/peer_test.go
@@ -23,7 +23,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/require"
+
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/p2p"
 
 	"github.com/status-im/status-go/waku/common"
 )
@@ -98,6 +101,26 @@ func TestPeerBasic(t *testing.T) {
 	}
 }
 
+func TestSendBundle(t *testing.T) {
+	rw1, rw2 := p2p.MsgPipe()
+	defer func() { handleError(t, rw1.Close()) }()
+	defer func() { handleError(t, rw2.Close()) }()
+	envelopes := []*common.Envelope{{
+		Expiry: 0,
+		TTL:    30,
+		Topic:  common.TopicType{1},
+		Data:   []byte{1, 1, 1},
+	}}
+
+	errc := make(chan error)
+	go func() {
+		_, err := sendBundle(rw1, envelopes)
+		errc <- err
+	}()
+	require.NoError(t, p2p.ExpectMsg(rw2, messagesCode, envelopes))
+	require.NoError(t, <-errc)
+}
+
 func generateMessageParams() (*common.MessageParams, error) {
 	// set all the parameters except p.Dst and p.Padding
 
@@ -122,4 +145,10 @@ func generateMessageParams() (*common.MessageParams, error) {
 	}
 
 	return &p, nil
+}
+
+func handleError(t *testing.T, err error) {
+	if err != nil {
+		t.Logf("deferred function error: '%s'", err)
+	}
 }
